+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Enhance [1.10.0] 2023-11-10
+- mission opening notification
+
 ## Enhance [1.9.0] 2023-10-24
 - show iframe in news guru